fix remaining non-strict TS errors
